Moves Statement debugging into ledger package and fixes tests
