+++ src/apps/chifra/pkg/types/types_statement.go
@@ -18,7 +18,6 @@ import (
 
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/base"
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/cache"
-	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/logger"
 )
 
 // EXISTING_CODE
@@ -667,95 +666,6 @@ func (s *Statement) IsStableCoin() bool {
 		usdt: true,
 	}
 	return stables[s.AssetAddr]
-}
-
-type LedgerContexter interface {
-	Prev() base.Blknum
-	Cur() base.Blknum
-	Next() base.Blknum
-	Recon() ReconType
-	Address() base.Address
-}
-
-func (s *Statement) DebugStatement(ctx LedgerContexter) {
-	logger.TestLog(true, "===================================================")
-	logger.TestLog(true, fmt.Sprintf("====> %s", s.AssetType))
-	logger.TestLog(true, "===================================================")
-	logger.TestLog(true, "Previous:              ", ctx.Prev())
-	logger.TestLog(true, "Current:               ", s.BlockNumber)
-	logger.TestLog(true, "Next:                  ", ctx.Next())
-	logger.TestLog(true, "reconciliationType:    ", s.ReconType.String())
-	logger.TestLog(true, "assetType:             ", s.AssetType)
-	logger.TestLog(true, "accountedFor:          ", s.AccountedFor)
-	logger.TestLog(true, "sender:                ", s.Sender, " ==> ", s.Recipient)
-	logger.TestLog(true, "assetAddr:             ", s.AssetAddr, "("+s.AssetSymbol+")", fmt.Sprintf("decimals: %d", s.Decimals))
-	logger.TestLog(true, "hash:                  ", s.TransactionHash)
-	logger.TestLog(true, "timestamp:             ", s.Timestamp)
-	if s.AssetType != TrialBalToken && s.AssetType != TrialBalNft {
-		logger.TestLog(true, fmt.Sprintf("blockNumber:            %d.%d", s.BlockNumber, s.TransactionIndex))
-	} else {
-		logger.TestLog(true, fmt.Sprintf("blockNumber:            %d.%d.%d", s.BlockNumber, s.TransactionIndex, s.LogIndex))
-	}
-	logger.TestLog(true, "priceSource:           ", s.SpotPrice, "("+s.PriceSource+")")
-	reportL("---------------------------------------------------")
-	logger.TestLog(true, "Trial balance:")
-	report1("   prevBal:            ", &s.PrevBal)
-	report2("   begBal:             ", &s.BegBal, s.EndBalDiff())
-	report1("   totalIn:            ", s.TotalIn())
-	report1("   totalOut:           ", s.TotalOut())
-	report1("   amountNet:          ", s.AmountNet())
-	reportL("                       =======================")
-	report2("   endBal:             ", &s.EndBal, s.BegBalDiff())
-	report1("   endBalCalc:         ", s.EndBalCalc())
-	reportL("---------------------------------------------------")
-	reportE("   amountIn:           ", &s.AmountIn)
-	reportE("   internalIn:         ", &s.InternalIn)
-	reportE("   minerBaseRewardIn:  ", &s.MinerBaseRewardIn)
-	reportE("   minerNephewRewardIn:", &s.MinerNephewRewardIn)
-	reportE("   minerTxFeeIn:       ", &s.MinerTxFeeIn)
-	reportE("   minerUncleRewardIn: ", &s.MinerUncleRewardIn)
-	reportE("   correctingIn:       ", &s.CorrectingIn)
-	reportE("   prefundIn:          ", &s.PrefundIn)
-	reportE("   selfDestructIn:     ", &s.SelfDestructIn)
-	reportE("   amountOut:          ", &s.AmountOut)
-	reportE("   internalOut:        ", &s.InternalOut)
-	reportE("   correctingOut:      ", &s.CorrectingOut)
-	reportE("   selfDestructOut:    ", &s.SelfDestructOut)
-	reportE("   gasOut:             ", &s.GasOut)
-	logger.TestLog(s.CorrectingReason != "", "   correctingReason:   ", s.CorrectingReason)
-	logger.TestLog(true, "   material:           ", s.IsMaterial())
-	logger.TestLog(true, "   reconciled:         ", s.Reconciled())
-	if !s.Reconciled() {
-		logger.TestLog(true, " ^^ we need to fix this ^^")
-	}
-	logger.TestLog(true, "---------------------------------------------------")
-	logger.TestLog(true, "End of trial balance report")
-}
-
-func reportE(msg string, val *base.Wei) {
-	isZero := func(val *base.Wei) bool {
-		return val.Cmp(base.NewWei(0)) == 0
-	}
-	logger.TestLog(!isZero(val), msg, val.ToEtherStr(18))
-}
-
-func report2(msg string, v1 *base.Wei, v2 *base.Wei) {
-	s := ""
-	if v1 != nil {
-		s = v1.ToEtherStr(18)
-	}
-	if v2 != nil {
-		s += " (" + v2.ToEtherStr(18) + ")"
-	}
-	logger.TestLog(true, msg, s)
-}
-
-func reportL(msg string) {
-	report2(msg, nil, nil)
-}
-
-func report1(msg string, val *base.Wei) {
-	report2(msg, val, nil)
 }
 
 // EXISTING_CODE
