+++ tests/gold/apps/acctExport/acctExport_accounting_to_cache.txt
@@ -48,6 +48,7 @@ TEST[DATE|TIME]    amountNet:           5
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              5 (0)
 TEST[DATE|TIME]    endBalCalc:          5
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    minerBaseRewardIn:   5
 TEST[DATE|TIME]    material:            true
@@ -84,6 +85,7 @@ TEST[DATE|TIME]    amountNet:           5
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              10 (0)
 TEST[DATE|TIME]    endBalCalc:          10
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    minerBaseRewardIn:   5
 TEST[DATE|TIME]    material:            true
@@ -121,6 +123,7 @@ TEST[DATE|TIME]    amountNet:           3.75
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              13.75 (0)
 TEST[DATE|TIME]    endBalCalc:          13.75
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    minerUncleRewardIn:  3.75
 TEST[DATE|TIME]    material:            true
@@ -159,6 +162,7 @@ TEST[DATE|TIME]    amountNet:           5
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              18.75 (0)
 TEST[DATE|TIME]    endBalCalc:          18.75
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    minerBaseRewardIn:   5
 TEST[DATE|TIME]    material:            true
@@ -198,6 +202,7 @@ TEST[DATE|TIME]    amountNet:           5
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              23.75 (0)
 TEST[DATE|TIME]    endBalCalc:          23.75
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    minerBaseRewardIn:   5
 TEST[DATE|TIME]    material:            true
@@ -238,6 +243,7 @@ TEST[DATE|TIME]    amountNet:           5
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              28.75 (0)
 TEST[DATE|TIME]    endBalCalc:          28.75
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    minerBaseRewardIn:   5
 TEST[DATE|TIME]    material:            true
@@ -279,6 +285,7 @@ TEST[DATE|TIME]    amountNet:           3.75
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              32.5 (0)
 TEST[DATE|TIME]    endBalCalc:          32.5
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    minerUncleRewardIn:  3.75
 TEST[DATE|TIME]    material:            true
@@ -321,6 +328,7 @@ TEST[DATE|TIME]    amountNet:           5
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              37.5 (0)
 TEST[DATE|TIME]    endBalCalc:          37.5
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    minerBaseRewardIn:   5
 TEST[DATE|TIME]    material:            true
@@ -364,6 +372,7 @@ TEST[DATE|TIME]    amountNet:           5
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              42.5 (0)
 TEST[DATE|TIME]    endBalCalc:          42.5
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    minerBaseRewardIn:   5
 TEST[DATE|TIME]    material:            true
@@ -408,6 +417,7 @@ TEST[DATE|TIME]    amountNet:           5.15625
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              47.65625 (0)
 TEST[DATE|TIME]    endBalCalc:          47.65625
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    minerBaseRewardIn:   5
 TEST[DATE|TIME]    minerNephewRewardIn: 0.15625
@@ -454,6 +464,7 @@ TEST[DATE|TIME]    amountNet:           5
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              52.65625 (0)
 TEST[DATE|TIME]    endBalCalc:          52.65625
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    minerBaseRewardIn:   5
 TEST[DATE|TIME]    material:            true
@@ -500,6 +511,7 @@ TEST[DATE|TIME]    amountNet:           5
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              57.65625 (0)
 TEST[DATE|TIME]    endBalCalc:          57.65625
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    minerBaseRewardIn:   5
 TEST[DATE|TIME]    material:            true
@@ -547,6 +559,7 @@ TEST[DATE|TIME]    amountNet:           5.15625
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              62.8125 (0)
 TEST[DATE|TIME]    endBalCalc:          62.8125
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    minerBaseRewardIn:   5
 TEST[DATE|TIME]    minerNephewRewardIn: 0.15625
