+++ tests/gold/apps/acctExport/acctExport_accounting_wei_1.txt
@@ -38,6 +38,7 @@ TEST[DATE|TIME]    amountNet:           4200
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              4200 (0)
 TEST[DATE|TIME]    endBalCalc:          4200
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    prefundIn:           4200
 TEST[DATE|TIME]    material:            true
@@ -74,6 +75,7 @@ TEST[DATE|TIME]    amountNet:           -4199.925320679510922664
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              0.074679320489077336 (0)
 TEST[DATE|TIME]    endBalCalc:          0.074679320489077336
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    amountOut:           4199.9
 TEST[DATE|TIME]    gasOut:              0.025320679510922664
@@ -112,6 +114,7 @@ TEST[DATE|TIME]    amountNet:           -0.074679320489077336
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              0 (0)
 TEST[DATE|TIME]    endBalCalc:          0
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    amountOut:           0.074259320489077336
 TEST[DATE|TIME]    gasOut:              0.00042
