+++ tests/gold/apps/acctExport/acctExport_statement_3.txt
@@ -39,6 +39,7 @@ TEST[DATE|TIME]    amountNet:           1970
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              1970 (0)
 TEST[DATE|TIME]    endBalCalc:          1970
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    prefundIn:           1970
 TEST[DATE|TIME]    material:            true
@@ -75,6 +76,7 @@ TEST[DATE|TIME]    amountNet:           -5.0105
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              1964.9895 (0)
 TEST[DATE|TIME]    endBalCalc:          1964.9895
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    amountOut:           5
 TEST[DATE|TIME]    gasOut:              0.0105
@@ -113,6 +115,7 @@ TEST[DATE|TIME]    amountNet:           -1964.9
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              0.0895 (0)
 TEST[DATE|TIME]    endBalCalc:          0.0895
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    amountOut:           1964.8895
 TEST[DATE|TIME]    gasOut:              0.0105
