+++ tests/gold/apps/acctExport/acctExport_statement_nft.txt
@@ -37,6 +37,7 @@ TEST[DATE|TIME]    amountNet:           0
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              4.9995342 (0)
 TEST[DATE|TIME]    endBalCalc:          4.9995342
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    material:            false
 TEST[DATE|TIME]    reconciled:          true
@@ -70,6 +71,7 @@ TEST[DATE|TIME]    amountNet:           22500
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              22500 (0)
 TEST[DATE|TIME]    endBalCalc:          22500
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    amountIn:            22500
 TEST[DATE|TIME]    material:            true
@@ -106,6 +108,7 @@ TEST[DATE|TIME]    amountNet:           0
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              4.9995342 (0)
 TEST[DATE|TIME]    endBalCalc:          4.9995342
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    material:            false
 TEST[DATE|TIME]    reconciled:          true
@@ -151,6 +154,7 @@ TEST[DATE|TIME]    amountNet:           0.000000000000000001
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              0.000000000000000001 (0)
 TEST[DATE|TIME]    endBalCalc:          0.000000000000000001
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    correctingOut:       -0.000000000000000001
 TEST[DATE|TIME]    correctingReason:    endbal
@@ -190,6 +194,7 @@ TEST[DATE|TIME]    amountNet:           0
 TEST[DATE|TIME]                        ======================= 
 TEST[DATE|TIME]    endBal:              4.9995342 (0)
 TEST[DATE|TIME]    endBalCalc:          4.9995342
+TEST[DATE|TIME]    rollingBalance:      0
 TEST[DATE|TIME] --------------------------------------------------- 
 TEST[DATE|TIME]    material:            false
 TEST[DATE|TIME]    reconciled:          true
