+++ tests/gold/apps/acctExport/sdk_tests/acctExport_accounting_tokens_1.txt
@@ -40,6 +40,7 @@ Trial balance:
                        ======================= 
    endBal:              4200 (0)
    endBalCalc:          4200
+   rollingBalance:      0
 --------------------------------------------------- 
    prefundIn:           4200
    material:            true
@@ -76,6 +77,7 @@ Trial balance:
                        ======================= 
    endBal:              0.074679320489077336 (0)
    endBalCalc:          0.074679320489077336
+   rollingBalance:      0
 --------------------------------------------------- 
    amountOut:           4199.9
    gasOut:              0.025320679510922664
@@ -114,6 +116,7 @@ Trial balance:
                        ======================= 
    endBal:              0 (0)
    endBalCalc:          0
+   rollingBalance:      0
 --------------------------------------------------- 
    amountOut:           0.074259320489077336
    gasOut:              0.00042
