+++ tests/gold/apps/acctExport/sdk_tests/acctExport_statement_2.txt
@@ -39,6 +39,7 @@ Trial balance:
                        ======================= 
    endBal:              1970 (0)
    endBalCalc:          1970
+   rollingBalance:      0
 --------------------------------------------------- 
    prefundIn:           1970
    material:            true
@@ -75,6 +76,7 @@ Trial balance:
                        ======================= 
    endBal:              1964.9895 (0)
    endBalCalc:          1964.9895
+   rollingBalance:      0
 --------------------------------------------------- 
    amountOut:           5
    gasOut:              0.0105
@@ -113,6 +115,7 @@ Trial balance:
                        ======================= 
    endBal:              0.0895 (0)
    endBalCalc:          0.0895
+   rollingBalance:      0
 --------------------------------------------------- 
    amountOut:           1964.8895
    gasOut:              0.0105
