+++ tests/gold/apps/acctExport/sdk_tests/acctExport_statement_filtered.txt
@@ -45,6 +45,7 @@ Trial balance:
                        ======================= 
    endBal:              7000 (0)
    endBalCalc:          7000
+   rollingBalance:      0
 --------------------------------------------------- 
    amountOut:           500
    material:            true
@@ -91,6 +92,7 @@ Trial balance:
                        ======================= 
    endBal:              0 (0)
    endBalCalc:          0
+   rollingBalance:      0
 --------------------------------------------------- 
    amountOut:           7000
    material:            true
