+++ tests/gold/apps/acctExport/sdk_tests/acctExport_statement_unfiltered.txt
@@ -53,6 +53,7 @@ Trial balance:
                        ======================= 
    endBal:              4.999709998 (0)
    endBalCalc:          4.999709998
+   rollingBalance:      0
 --------------------------------------------------- 
    gasOut:              0.000034017
    material:            true
@@ -89,6 +90,7 @@ Trial balance:
                        ======================= 
    endBal:              7000 (0)
    endBalCalc:          7000
+   rollingBalance:      0
 --------------------------------------------------- 
    amountOut:           500
    material:            true
@@ -138,6 +140,7 @@ Trial balance:
                        ======================= 
    endBal:              4.999663949 (0)
    endBalCalc:          4.999663949
+   rollingBalance:      0
 --------------------------------------------------- 
    gasOut:              0.000046049
    material:            true
@@ -189,6 +192,7 @@ Trial balance:
                        ======================= 
    endBal:              4.999639408 (0)
    endBalCalc:          4.999639408
+   rollingBalance:      0
 --------------------------------------------------- 
    gasOut:              0.000024541
    material:            true
@@ -225,6 +229,7 @@ Trial balance:
                        ======================= 
    endBal:              0 (0)
    endBalCalc:          0
+   rollingBalance:      0
 --------------------------------------------------- 
    amountOut:           7000
    material:            true
